byebye warnings

+++ Userland/SampleCodeModule/stdlib.c
@@ -190,12 +190,12 @@ void * malloc(uint64_t size) {
     if (size == 0) return NULL;
     void *ptr = syscall_allocMemory(size);
     if (ptr == NULL) {
-        printferror("Error allocating memory of size %llu\n", size);
+        printferror("Error allocating memory of size %d\n", size);
     }
     return ptr;
 }
 
 void free(void *ptr) {
     if (ptr == NULL) return;
-    syscall_freeMemory((uint64_t)ptr);
+    syscall_freeMemory(ptr);
 }
